% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{create_connection_object}
\alias{create_connection_object}
\title{Create a sharekhanconnect connection object.}
\usage{
create_connection_object(params)
}
\arguments{
\item{params}{a list mapping object property name to values. See details
for more.}
}
\value{
S4 object for sharekhanconnect which can be used in further api calls.
}
\description{
This function returns an object of sharekhanconnect type, given the
input parametrs.
}
\details{
This function creates an S4 object from the input parameters. The
required parameters are api_key, api_secret. If root (the api root), routes
(the list of api endpoints) and login (the login url) is supplied, they are
used. Otherwise the internally defined values are used. Only a single
instance of this class per \code{api_key} should be initialized.
Parameters should be a named list of all inputs to create a connection.
The important attributes are: api_key( character) - your api key;
api_secret(character) - your api secret (both of these are available on
the developer page in your sharekhan app, and must be supplied during object
creation); request_token(character) - should be set after a successful
login flow; access_token(character) - obtained using fetch_access_token
}
