% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restful.R
\name{get_funds}
\alias{get_funds}
\title{Function to get the fund details.}
\usage{
get_funds(object, param)
}
\arguments{
\item{object}{An object of type sharekhanconnect with valid api_key and
access_token.}

\item{param}{This contains the request parameters.}
}
\value{
Returns a json object with fund details, if successful.
}
\description{
Gets available cash limit for different segments.
}
