% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restful.R
\name{get_historical}
\alias{get_historical}
\title{Function to get the historical data upto last traded date.}
\usage{
get_historical(object, param)
}
\arguments{
\item{object}{An object of type sharekhanconnect with valid api_key and
access_token.}

\item{param}{This contains the request parameters.}
}
\value{
Returns a json object with historical data, if successful.
}
\description{
Gets historical data upto last traded date. The candle data is
based on minutes. The highest timeframe is available upto 90 minutes.
}
