% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restful.R
\name{get_holdings}
\alias{get_holdings}
\title{Function to get the holdings details.}
\usage{
get_holdings(object, param)
}
\arguments{
\item{object}{An object of type sharekhanconnect with valid api_key and
access_token.}

\item{param}{This contains the request parameters.}
}
\value{
Returns a json object with holdings details, if successful.
}
\description{
It contains user's long term stocks.
}
