% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{get_login_url}
\alias{get_login_url}
\title{Function to get the login URL.}
\usage{
get_login_url(object, version_id = NULL)
}
\arguments{
\item{object}{An object of sharekhanconnect type}

\item{vendor_key}{A vendor_key with default value NULL}
}
\value{
Returns the login URL.
}
\description{
Function to get the login URL (the login entry + api_key) from
the sharekhanconnect object.
}
