% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restful.R
\name{get_master}
\alias{get_master}
\title{Function to get the active scrips for the day.}
\usage{
get_master(object, param)
}
\arguments{
\item{object}{An object of type sharekhanconnect with valid api_key and
access_token.}

\item{param}{This contains the request parameters.}
}
\value{
Returns a json object with scrips details, if successful.
}
\description{
Gets active scrips for the day.
}
