% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restful.R
\name{get_report_trade}
\alias{get_report_trade}
\title{Function to get the trade history.}
\usage{
get_report_trade(object, param)
}
\arguments{
\item{object}{An object of type sharekhanconnect with valid api_key and
access_token.}

\item{param}{This contains the request parameters.}
}
\value{
Returns a json object with trade details, if successful.
}
\description{
Gets trade history.
}
