% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restful.R
\name{get_trade_position}
\alias{get_trade_position}
\title{Function to get the trades details.}
\usage{
get_trade_position(object, param)
}
\arguments{
\item{object}{An object of type sharekhanconnect with valid api_key and
access_token.}

\item{param}{This contains the request parameters.}
}
\value{
Returns a json object with trades details, if successful.
}
\description{
Gets trades history.
}
