% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restful.R
\name{place_order}
\alias{place_order}
\title{Function to place a new order.}
\usage{
place_order(object, param)
}
\arguments{
\item{object}{An object of type sharekhanconnect with valid api_key and
access_token.}

\item{param}{This contains the request parameters.}
}
\value{
Returns a json object with order details, if successful.
}
\description{
Function to place a order with specified details.
}
