% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{set_access_token}
\alias{set_access_token}
\title{Function to set the access token in the sharekhanconnect object.}
\usage{
set_access_token(object, access_token)
}
\arguments{
\item{object}{a sharekhanconnect object.}

\item{access_token}{Access token.}
}
\value{
sharekhan connect object with access token set.
}
\description{
This function sets the access token (obtained by sending a
hash of api key, request token and api secret) inside the sharekhanconnect
object.
}
\seealso{
\code{\link[sharekhanconnect]{create_connection_object}}

\code{\link[sharekhanconnect]{fetch_access_token}}
}
